Even more cleanup

+++ src/cpp/server/load_reporter/load_data_store.h
@@ -74,6 +74,7 @@ class LoadRecordKey {
         user_id_(std::move(user_id)),
         client_ip_hex_(std::move(client_ip_hex)) {}
 
+  // Parses the input client_ip_and_token to set client IP, LB ID, and LB tag.
   explicit LoadRecordKey(const grpc::string& client_ip_and_token,
                          grpc::string user_id);
 
@@ -88,7 +89,7 @@ class LoadRecordKey {
            user_id_ == other.user_id_ && client_ip_hex_ == other.client_ip_hex_;
   }
 
-  // Get the client IP bytes in network order (i.e., big-endian).
+  // Gets the client IP bytes in network order (i.e., big-endian).
   grpc::string GetClientIpBytes() const;
 
   // Getters.
@@ -123,8 +124,8 @@ class LoadRecordKey {
 // The value of a load record.
 class LoadRecordValue {
  public:
-  explicit LoadRecordValue(uint64_t start_count = 0, uint64_t ok_count = 0,
-                           uint64_t error_count = 0, double bytes_sent = 0,
+  explicit LoadRecordValue(double start_count = 0, double ok_count = 0,
+                           double error_count = 0, double bytes_sent = 0,
                            double bytes_recv = 0, double latency_ms = 0)
       : start_count_(start_count),
         ok_count_(ok_count),
@@ -169,9 +170,9 @@ class LoadRecordValue {
   }
 
   // Getters.
-  uint64_t start_count() const { return start_count_; }
-  uint64_t ok_count() const { return ok_count_; }
-  uint64_t error_count() const { return error_count_; }
+  double start_count() const { return start_count_; }
+  double ok_count() const { return ok_count_; }
+  double error_count() const { return error_count_; }
   double bytes_sent() const { return bytes_sent_; }
   double bytes_recv() const { return bytes_recv_; }
   double latency_ms() const { return latency_ms_; }
@@ -181,9 +182,9 @@ class LoadRecordValue {
   }
 
  private:
-  uint64_t start_count_ = 0;
-  uint64_t ok_count_ = 0;
-  uint64_t error_count_ = 0;
+  double start_count_ = 0;
+  double ok_count_ = 0;
+  double error_count_ = 0;
   double bytes_sent_ = 0;
   double bytes_recv_ = 0;
   double latency_ms_ = 0;
